requirements.txt
asgiref==3.7.2
CacheControl==0.12.14
cachetools==5.3.1
cachy==0.3.0
cleo==0.8.1
clikit==0.6.2
contourpy==1.1.0
crashtest==0.3.1
cycler==0.11.0
distlib==0.3.7
Django==4.2.4
django-realtime==1.1
django-tables==0.2
django-tables2==2.6.0
electionguard==1.2.3
et-xmlfile==1.1.0
filelock==3.12.2
fonttools==4.42.1
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
html5lib==1.1
httplib2==0.22.0
jeepney==0.8.0
keyring==21.8.0
kiwisolver==1.4.5
lockfile==0.12.2
matplotlib==3.7.2
msgpack==1.0.5
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==20.9
pandas==2.0.3
pastel==0.2.1
pexpect==4.8.0
Pillow==10.0.0
platformdirs==3.10.0
poetry==1.1.6
poetry-core==1.0.8
protobuf==4.24.2
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pylev==1.4.0
python-dateutil==2.8.2
pytz==2023.3
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
shellingham==1.5.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
tzdata==2023.3
uritemplate==4.1.1
virtualenv==20.24.3
webencodings==0.5.1
xlrd==2.0.1
zipp==3.6.0
